Moved engine to separated package, fixed dependencies issues
