+++ shell/context.go
@@ -2,6 +2,7 @@ package shell
 
 import (
 	"context"
+	"github.com/gruntwork-io/terragrunt/util"
 
 	"github.com/gruntwork-io/terragrunt/options"
 )
@@ -10,7 +11,7 @@ const TerraformCommandContextKey ctxKey = iota
 
 type ctxKey byte
 
-type RunShellCommandFunc func(ctx context.Context, opts *options.TerragruntOptions, args []string) (*CmdOutput, error)
+type RunShellCommandFunc func(ctx context.Context, opts *options.TerragruntOptions, args []string) (*util.CmdOutput, error)
 
 func ContextWithTerraformCommandHook(ctx context.Context, fn RunShellCommandFunc) context.Context {
 	return context.WithValue(ctx, TerraformCommandContextKey, fn)
